+++ provision/host.sh
@@ -617,7 +617,7 @@ configure_etc_hosts()
 
 	for _j in $JAIL_ORDERED_LIST; do
 		_hosts="$_hosts
-$(get_jail_ip "$_j")		$_j"
+$(get_jail_ip4 "$_j")		$_j"
 	done
 
 	echo "$_hosts" >> "/etc/hosts"